requirements.txt
tqdm
requests
mysql-connector
python-dotenv